requirements.txt
glymur==0.14.3
h5py==3.11.0
matplotlib==3.10.3
nibabel==5.3.2
numpy==2.3.1
pandas==2.3.0
Pillow==11.2.1
pydicom==3.0.1
PyYAML==6.0.2
PyYAML==6.0.2
scipy==1.16.0
skimage==0.0
torch==2.4.1
torchmetrics==1.4.2
torchmetrics==1.7.1
torchvision==0.19.1
tqdm==4.66.5
wandb==0.20.1
